do not filter invalid locale resources from the built assets
