+++ src/Tasks/Microsoft.NET.Build.Tasks.UnitTests/GivenAAllowEmptyTelemetry.cs
@@ -131,55 +131,6 @@ namespace Microsoft.NET.Build.Tasks.UnitTests
                     _log.AppendLine(message);
                 }
             }
-
-            public bool BuildProjectFile(string projectFileName, string[] targetNames, IDictionary globalProperties,
-                IDictionary targetOutputs, string toolsVersion)
-            {
-                throw new NotImplementedException();
-            }
-
-            public bool BuildProjectFilesInParallel(string[] projectFileNames, string[] targetNames,
-                IDictionary[] globalProperties,
-                IDictionary[] targetOutputsPerProject, string[] toolsVersion, bool useResultsCache,
-                bool unloadProjectsOnCompletion)
-            {
-                throw new NotImplementedException();
-            }
-
-            public bool IsRunningMultipleNodes { get; }
-
-            public BuildEngineResult BuildProjectFilesInParallel(string[] projectFileNames, string[] targetNames,
-                IDictionary[] globalProperties, IList<string>[] removeGlobalProperties, string[] toolsVersion,
-                bool returnTargetOutputs)
-            {
-                throw new NotImplementedException();
-            }
-
-            public void Yield()
-            {
-                throw new NotImplementedException();
-            }
-
-            public void Reacquire()
-            {
-                throw new NotImplementedException();
-            }
-
-            public void RegisterTaskObject(object key, object obj, RegisteredTaskObjectLifetime lifetime,
-                bool allowEarlyCollection)
-            {
-                throw new NotImplementedException();
-            }
-
-            public object GetRegisteredTaskObject(object key, RegisteredTaskObjectLifetime lifetime)
-            {
-                throw new NotImplementedException();
-            }
-
-            public object UnregisterTaskObject(object key, RegisteredTaskObjectLifetime lifetime)
-            {
-                throw new NotImplementedException();
-            }
         }
     }
 }
