+++ src/Tasks/Microsoft.NET.Build.Tasks.UnitTests/MockNeverCacheBuildEngine4.cs
@@ -8,53 +8,11 @@ using Microsoft.Build.Framework;
 
 namespace Microsoft.NET.Build.Tasks.UnitTests
 {
-    internal class MockNeverCacheBuildEngine4 : MockBuildEngine, IBuildEngine4
+    internal class MockNeverCacheBuildEngine4 : MockBuildEngine
     {
-        public bool IsRunningMultipleNodes => throw new NotImplementedException();
-
-        public bool BuildProjectFile(string projectFileName, string[] targetNames, IDictionary globalProperties,
-            IDictionary targetOutputs, string toolsVersion)
-        {
-            throw new NotImplementedException();
-        }
-
-        public BuildEngineResult BuildProjectFilesInParallel(string[] projectFileNames, string[] targetNames,
-            IDictionary[] globalProperties, IList<string>[] removeGlobalProperties, string[] toolsVersion,
-            bool returnTargetOutputs)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool BuildProjectFilesInParallel(string[] projectFileNames, string[] targetNames,
-            IDictionary[] globalProperties, IDictionary[] targetOutputsPerProject, string[] toolsVersion,
-            bool useResultsCache, bool unloadProjectsOnCompletion)
-        {
-            throw new NotImplementedException();
-        }
-
-        public object GetRegisteredTaskObject(object key, RegisteredTaskObjectLifetime lifetime)
+        public override object GetRegisteredTaskObject(object key, RegisteredTaskObjectLifetime lifetime)
         {
             return null;
-        }
-
-        public void Reacquire()
-        {
-            throw new NotImplementedException();
-        }
-
-        public void RegisterTaskObject(object key, object obj, RegisteredTaskObjectLifetime lifetime,
-            bool allowEarlyCollection)
-        {
-        }
-
-        public object UnregisterTaskObject(object key, RegisteredTaskObjectLifetime lifetime)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void Yield()
-        {
-            throw new NotImplementedException();
         }
     }
 }
