+++ src/Tasks/Microsoft.NET.Build.Tasks.UnitTests/Mocks/MockBuildEngine.cs
@@ -8,7 +8,7 @@ using Microsoft.Build.Framework;
 
 namespace Microsoft.NET.Build.Tasks.UnitTests
 {
-    public class MockBuildEngine : IBuildEngine
+    public class MockBuildEngine : IBuildEngine4
     {
         public int ColumnNumberOfTaskNode { get; set; }
 
@@ -43,9 +43,35 @@ namespace Microsoft.NET.Build.Tasks.UnitTests
             Warnings.Add(e);
         }
 
+        public virtual object GetRegisteredTaskObject(object key, RegisteredTaskObjectLifetime lifetime)
+        {
+            RegisteredTaskObjectsQueries++;
+
+            RegisteredTaskObjects.TryGetValue(key, out object ret);
+            return ret;
+        }
+        
+        public void RegisterTaskObject(object key, object obj, RegisteredTaskObjectLifetime lifetime, bool allowEarlyCollection)
+        {
+            RegisteredTaskObjects.Add(key, obj);
+        }
+
+        public object UnregisterTaskObject(object key, RegisteredTaskObjectLifetime lifetime) => null;
+        public BuildEngineResult BuildProjectFilesInParallel(string[] projectFileNames, string[] targetNames, IDictionary[] globalProperties, IList<string>[] removeGlobalProperties, string[] toolsVersion, bool returnTargetOutputs) => new BuildEngineResult();
+        public void Yield() {
+        }
+        public void Reacquire() {
+        }
+        public bool BuildProjectFile(string projectFileName, string[] targetNames, IDictionary globalProperties, IDictionary targetOutputs, string toolsVersion) => false;
+        public bool BuildProjectFilesInParallel(string[] projectFileNames, string[] targetNames, IDictionary[] globalProperties, IDictionary[] targetOutputsPerProject, string[] toolsVersion, bool useResultsCache, bool unloadProjectsOnCompletion) => false;
+
         public IList<CustomBuildEventArgs> CustomEvents { get; } = new List<CustomBuildEventArgs>();
         public IList<BuildErrorEventArgs> Errors { get; } = new List<BuildErrorEventArgs>();
         public IList<BuildMessageEventArgs> Messages { get; } = new List<BuildMessageEventArgs>();
         public IList<BuildWarningEventArgs> Warnings { get; } = new List<BuildWarningEventArgs>();
+        public Dictionary<object, object> RegisteredTaskObjects { get; } = new Dictionary<object, object>();
+        public int RegisteredTaskObjectsQueries = 0;
+
+        public bool IsRunningMultipleNodes => throw new NotImplementedException();
     }
 }
